% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANPRS.R
\name{debug_test_pkg}
\alias{debug_test_pkg}
\title{Run the package on the provided data set
DEBUG VERSION}
\usage{
debug_test_pkg(
  summaryZ,
  funcIndex,
  debug_output = FALSE,
  sub_tuning = 1,
  lambda_vec_limit_len = c(1.5, 1)
)
}
\description{
Run the package on the provided data set
DEBUG VERSION
}
