% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsfPEN_cpp}
\alias{gsfPEN_cpp}
\title{Main CPP function}
\usage{
gsfPEN_cpp(
  summary_betas,
  ld_J,
  num_iter_vec,
  index_matrix,
  index_J,
  ld_vec,
  upper_val,
  SD_vec,
  tuning_matrix,
  beta_matrix,
  lambda0_vec,
  z_matrix,
  all_tuning_matrix,
  lambda_vec_func,
  func_lambda,
  Ifunc_SNP,
  dims,
  params
)
}
\arguments{
\item{summary_betas}{matrix}

\item{ld_J}{vector}

\item{num_iter_vec}{vector}

\item{index_matrix}{matrix}

\item{index_J}{vector}

\item{ld_vec}{vector}

\item{upper_val}{double}

\item{SD_vec}{matrix}

\item{tuning_matrix}{matrix}

\item{beta_matrix}{matrix}

\item{lambda0_vec}{vector}

\item{z_matrix}{matrix}

\item{all_tuning_matrix}{matrix}

\item{lambda_vec_func}{vector}

\item{func_lambda}{matrix}

\item{Ifunc_SNP}{vector}

\item{dims}{vector}

\item{params}{vector}
}
\description{
Main CPP function
}
