% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANPRS.R
\name{gsfPEN_R}
\alias{gsfPEN_R}
\title{Main Function}
\usage{
gsfPEN_R(
  summary_z,
  n_vec,
  plinkLD,
  func_index,
  n_iter = 1000,
  upper_val = NULL,
  breaking = 1,
  z_scale = 1,
  tuning_matrix = NULL,
  p_threshold = NULL,
  p_threshold_params = c(0.5, 10^-4, 4),
  tau_factor = c(1/25, 1, 3),
  len_lim_lambda = 4,
  sub_tuning = 4,
  lim_lambda = c(0.5, 0.9),
  len_lambda = 4,
  lambda_vec = NULL,
  lambda_vec_limit_len = c(1.5, 3),
  df_max = NULL,
  debug_output = FALSE
)
}
\description{
Main Function
}
