% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsfPEN_sparse_cpp}
\alias{gsfPEN_sparse_cpp}
\title{Main CPP function}
\usage{
gsfPEN_sparse_cpp(
  summary_betas,
  ld_J,
  index_matrix,
  index_J,
  ld_vec,
  SD_vec,
  tuning_matrix,
  lambda0_vec,
  z_matrix,
  lambda_vec_func,
  func_lambda,
  Ifunc_SNP,
  dims,
  params
)
}
\arguments{
\item{summary_betas}{matrix of summary statistics}

\item{ld_J}{vector of indices of SNPs in LD with the current SNP}

\item{index_matrix}{matrix of indices of SNPs in LD with the current SNP}

\item{index_J}{vector of indices of SNPs in LD with the current SNP}

\item{ld_vec}{vector of LD values}

\item{SD_vec}{matrix of SD values}

\item{tuning_matrix}{matrix of tuning parameters}

\item{lambda0_vec}{vector of lambda0 values}

\item{z_matrix}{matrix of z values}

\item{lambda_vec_func}{vector of lambda values}

\item{func_lambda}{matrix of lambda values}

\item{Ifunc_SNP}{vector of indices of SNPs in LD with the current SNP}

\item{dims}{vector of dimensions}

\item{params}{vector of parameters}
}
\description{
Main CPP function
}
