% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsPEN_sparse_cpp}
\alias{gsPEN_sparse_cpp}
\title{Main CPP function}
\usage{
gsPEN_sparse_cpp(
  summary_betas,
  ld_J,
  index_matrix,
  index_J,
  ld_vec,
  SD_vec,
  tuning_matrix,
  dims,
  params
)
}
\arguments{
\item{summary_betas}{matrix of summary statistics}

\item{ld_J}{vector of indices of SNPs in LD with the current SNP}

\item{index_matrix}{matrix of indices of SNPs in LD with the current SNP}

\item{index_J}{vector of indices of SNPs in LD with the current SNP}

\item{ld_vec}{vector of LD values}

\item{SD_vec}{matrix of SD values}

\item{tuning_matrix}{matrix of tuning parameters}

\item{dims}{vector of dimensions}

\item{params}{vector of parameters}
}
\description{
Main CPP function
}
