% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_gsPEN}
\alias{test_gsPEN}
\title{Run gsPEN on the provided data set}
\usage{
test_gsPEN(...)
}
\arguments{
\item{...}{Additional arguments to pass to gsPEN_R}
}
\description{
Run gsPEN on the provided data set
}
