% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_gsfPEN}
\alias{test_gsfPEN}
\title{Run gsfPEN on a small sample of the provided data set (Only 100 samples)}
\usage{
test_gsfPEN(...)
}
\arguments{
\item{...}{Additional arguments to pass to gsfPEN_R}
}
\value{
The output of gsfPEN_R
}
\description{
Run gsfPEN on a small sample of the provided data set (Only 100 samples)
}
