% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_gsfPEN}
\alias{test_gsfPEN}
\title{Run gsfPEN on the provided data set}
\usage{
test_gsfPEN(...)
}
\arguments{
\item{...}{Additional arguments to pass to gsfPEN_R}
}
\description{
Run gsfPEN on the provided data set
}
