% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{test_gsPEN}
\alias{test_gsPEN}
\title{Run gsPEN on a small sample of the provided data set (Only 100 samples)}
\usage{
test_gsPEN(...)
}
\arguments{
\item{...}{Additional arguments to pass to gsPEN_R}
}
\value{
The output of gsPEN_R
}
\description{
Run gsPEN on a small sample of the provided data set (Only 100 samples)
}
